Replace old Varda integration logic
Split new functionality into a new client

Mainly extract throwing functions for each operation

Remove old client and dead code

Add minor client tweaks

Remove V2 suffixes
