+++ service/src/main/kotlin/fi/espoo/evaka/varda/integration/VardaClient.kt
@@ -4,7 +4,6 @@
 
 package fi.espoo.evaka.varda.integration
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
@@ -19,14 +18,12 @@ import com.github.kittinunf.result.Result
 import fi.espoo.evaka.VardaEnv
 import fi.espoo.evaka.shared.utils.responseStringWithRetries
 import fi.espoo.evaka.shared.utils.token
-import fi.espoo.evaka.varda.VardaChildPayload
 import fi.espoo.evaka.varda.VardaChildRequest
 import fi.espoo.evaka.varda.VardaChildResponse
 import fi.espoo.evaka.varda.VardaDecision
 import fi.espoo.evaka.varda.VardaDecisionResponse
 import fi.espoo.evaka.varda.VardaFeeData
 import fi.espoo.evaka.varda.VardaFeeDataResponse
-import fi.espoo.evaka.varda.VardaPaosChildPayload
 import fi.espoo.evaka.varda.VardaPersonRequest
 import fi.espoo.evaka.varda.VardaPersonResponse
 import fi.espoo.evaka.varda.VardaPlacement
@@ -36,8 +33,6 @@ import fi.espoo.evaka.varda.VardaUnitResponse
 import fi.espoo.evaka.varda.VardaUpdateOrganizer
 import fi.espoo.voltti.logging.loggers.error
 import mu.KotlinLogging
-import java.time.LocalDate
-import java.util.UUID
 
 private val logger = KotlinLogging.logger {}
 
@@ -62,9 +57,8 @@ class VardaClient(
     val getPlacementUrl = { placementId: Long -> "$placementUrl$placementId/" }
     val sourceSystem: String = env.sourceSystem
 
-    // Varda error may contain any of the keys below, perhaps, depending on the endpoint
     private data class VardaErrors(
-        val errors: List<VardaError> = listOf(),
+        val errors: List<VardaError>,
         val tuntimaara_viikossa: List<VardaError> = listOf(),
         val huoltajat: List<VardaError> = listOf()
     ) {
@@ -120,89 +114,10 @@ class VardaClient(
 
     private fun vardaError(request: Request, error: FuelError, message: (meta: VardaRequestError) -> String): Nothing {
         val meta = parseVardaError(request, error)
-        logger.error(request, meta.asMap()) { "VardaUpdate: request failed to ${meta.url}, status ${meta.statusCode}, reason ${meta.errorCode}: ${meta.errorMessage}" }
+        logger.error(request, meta.asMap()) {
+            "VardaUpdate: request failed to ${meta.url}, status ${meta.statusCode}, reason ${meta.errorCode}: ${meta.errorMessage}"
+        }
         error(message(meta))
-    }
-
-    fun createUnit(unit: VardaUnitRequest): VardaUnitResponse? {
-        logger.info { "Creating a new unit ${unit.nimi} to Varda" }
-        val (request, _, result) = fuel.post(unitUrl)
-            .jsonBody(objectMapper.writeValueAsString(unit))
-            .authenticatedResponseStringWithRetries()
-
-        return when (result) {
-            is Result.Success -> {
-                logger.info { "Creating a new unit ${unit.nimi} to Varda succeeded" }
-                objectMapper.readValue<VardaUnitResponse>(
-                    objectMapper.readTree(result.get()).toString()
-                )
-            }
-            is Result.Failure -> {
-                logRequestError(request, result.error)
-                null
-            }
-        }
-    }
-
-    fun updateUnit(unit: VardaUnitRequest): VardaUnitResponse? {
-        logger.info { "Updating unit ${unit.nimi} to Varda" }
-        val url = "$unitUrl${unit.id}/"
-        val (request, _, result) = fuel.put(url)
-            .jsonBody(objectMapper.writeValueAsString(unit))
-            .authenticatedResponseStringWithRetries()
-
-        return when (result) {
-            is Result.Success -> {
-                logger.info { "Updating unit ${unit.nimi} to Varda succeeded" }
-                objectMapper.readValue<VardaUnitResponse>(
-                    objectMapper.readTree(result.get()).toString()
-                )
-            }
-            is Result.Failure -> {
-                logRequestError(request, result.error)
-                null
-            }
-        }
-    }
-
-    fun updateOrganizer(organizer: VardaUpdateOrganizer): Boolean {
-        logger.info { "Updating organizer to Varda" }
-
-        val (request, _, result) = fuel.put("$organizerUrl${organizer.vardaOrganizerId}")
-            .header(Headers.CONTENT_TYPE, "application/json")
-            .jsonBody(objectMapper.writeValueAsString(organizer))
-            .authenticatedResponseStringWithRetries()
-
-        return when (result) {
-            is Result.Success -> {
-                logger.info { "Updating organizer to Varda succeeded" }
-                true
-            }
-            is Result.Failure -> {
-                logRequestError(request, result.error)
-                false
-            }
-        }
-    }
-
-    fun createPerson(newPerson: VardaPersonRequest): VardaPersonResponse? {
-        logger.info { "Creating a new person ${newPerson.id} to Varda" }
-
-        val (request, _, result) = fuel.post(personUrl)
-            .jsonBody(objectMapper.writeValueAsString(newPerson)).authenticatedResponseStringWithRetries()
-
-        return when (result) {
-            is Result.Success -> {
-                logger.info { "Creating a new person ${newPerson.id} to Varda succeeded" }
-                objectMapper.readValue<VardaPersonResponse>(
-                    objectMapper.readTree(result.get()).toString()
-                )
-            }
-            is Result.Failure -> {
-                logRequestError(request, result.error)
-                null
-            }
-        }
     }
 
     data class VardaPersonSearchRequest(
@@ -239,27 +154,26 @@ class VardaClient(
         }
     }
 
-    fun createChild(child: VardaChildRequest): VardaChildResponse? {
-        logger.info { "Creating child to Varda (body: $child)" }
+    fun createPerson(person: VardaPersonRequest): VardaPersonResponse {
+        logger.info { "VardaUpdate: client sending person (body: $person)" }
 
-        val (request, _, result) = fuel.post(childUrl)
-            .jsonBody(objectMapper.writeValueAsString(child)).authenticatedResponseStringWithRetries()
+        val (request, _, result) = fuel.post(personUrl)
+            .jsonBody(objectMapper.writeValueAsString(person)).authenticatedResponseStringWithRetries()
 
         return when (result) {
             is Result.Success -> {
-                logger.info { "Creating child to Varda succeeded (body: $child)" }
-                objectMapper.readValue<VardaChildResponse>(
-                    objectMapper.readTree(result.get()).toString()
-                )
+                logger.info { "VardaUpdate: client successfully sent child (body: $person)" }
+                objectMapper.readValue(result.get())
             }
             is Result.Failure -> {
-                logRequestError(request, result.error)
-                null
+                vardaError(request, result.error) { err ->
+                    "VardaUpdate: client failed to send person: $err"
+                }
             }
         }
     }
 
-    fun createChildV2(child: VardaChildRequest): VardaChildResponse {
+    fun createChild(child: VardaChildRequest): VardaChildResponse {
         logger.info { "VardaUpdate: client sending child (body: $child)" }
 
         val (request, _, result) = fuel.post(childUrl)
@@ -278,116 +192,7 @@ class VardaClient(
         }
     }
 
-    fun createFeeData(feeData: VardaFeeData): VardaFeeDataResponse? {
-        logger.info { "Creating fee data for child ${feeData.lapsi} to Varda" }
-
-        val (request, _, result) = fuel.post(feeDataUrl)
-            .jsonBody(objectMapper.writeValueAsString(feeData)).authenticatedResponseStringWithRetries()
-
-        return when (result) {
-            is Result.Success -> {
-                logger.info { "Creating fee data for child ${feeData.lapsi} to Varda succeeded" }
-                objectMapper.readValue<VardaFeeDataResponse>(
-                    objectMapper.readTree(result.get()).toString()
-                )
-            }
-            is Result.Failure -> {
-                logRequestError(request, result.error)
-                null
-            }
-        }
-    }
-
-    fun createFeeDataV2(feeData: VardaFeeData): VardaFeeDataResponse {
-        logger.info { "VardaUpdate: client sending fee data for child ${feeData.lapsi}" }
-        val (request, _, result) = fuel.post(feeDataUrl)
-            .jsonBody(objectMapper.writeValueAsString(feeData)).authenticatedResponseStringWithRetries()
-
-        return when (result) {
-            is Result.Success -> {
-                logger.info { "VardaUpdate: client successfully sent fee data for child ${feeData.lapsi}" }
-                objectMapper.readValue(result.get())
-            }
-            is Result.Failure -> {
-                vardaError(request, result.error) { err ->
-                    "VardaUpdate: client failed to send fee data for child ${feeData.lapsi}: ${err.errorCode}, $err"
-                }
-            }
-        }
-    }
-
-    fun updateFeeData(vardaFeeDataId: Long, feeData: VardaFeeData): Boolean {
-        logger.info { "Updating fee data $vardaFeeDataId to Varda" }
-
-        val (request, _, result) = fuel.put("$feeDataUrl$vardaFeeDataId")
-            .jsonBody(objectMapper.writeValueAsString(feeData)).authenticatedResponseStringWithRetries()
-
-        return when (result) {
-            is Result.Success -> {
-                logger.info { "Updating fee data $vardaFeeDataId to Varda succeeded" }
-                true
-            }
-            is Result.Failure -> {
-                logRequestError(request, result.error)
-                false
-            }
-        }
-    }
-
-    fun deleteFeeData(vardaId: Long): Boolean {
-        logger.info { "Deleting fee data $vardaId from Varda" }
-
-        val (request, _, result) = fuel.delete("$feeDataUrl$vardaId/").authenticatedResponseStringWithRetries()
-
-        return when (result) {
-            is Result.Success -> {
-                logger.info { "Deleting fee data $vardaId from Varda succeeded" }
-                true
-            }
-            is Result.Failure -> {
-                logRequestError(request, result.error)
-                false
-            }
-        }
-    }
-
-    fun deleteFeeDataV2(vardaId: Long): Boolean {
-        logger.info { "VardaUpdate: client deleting fee data $vardaId" }
-
-        val (request, _, result) = fuel.delete("$feeDataUrl$vardaId/").authenticatedResponseStringWithRetries()
-
-        return when (result) {
-            is Result.Success -> {
-                logger.info { "VardaUpdate: client successfully deleted fee data $vardaId" }
-                true
-            }
-            is Result.Failure -> {
-                vardaError(request, result.error) { err ->
-                    "VardaUpdate: client failed to delete fee data $vardaId: $err"
-                }
-            }
-        }
-    }
-
-    fun createDecision(newDecision: VardaDecision): VardaDecisionResponse? {
-        logger.info { "Creating a new decision to Varda (body: $newDecision)" }
-
-        val (request, _, result) = fuel.post(decisionUrl)
-            .jsonBody(objectMapper.writeValueAsString(newDecision)).authenticatedResponseStringWithRetries()
-
-        return when (result) {
-            is Result.Success -> {
-                logger.info { "Creating a new decision to Varda succeeded (body: $newDecision)" }
-                objectMapper.readValue(result.get())
-            }
-            is Result.Failure -> {
-                logRequestError(request, result.error)
-                null
-            }
-        }
-    }
-
-    fun createDecisionV2(newDecision: VardaDecision): VardaDecisionResponse {
+    fun createDecision(newDecision: VardaDecision): VardaDecisionResponse {
         logger.info { "VardaUpdate: client sending new decision (body: $newDecision)" }
         val (request, _, result) = fuel.post(decisionUrl)
             .jsonBody(objectMapper.writeValueAsString(newDecision)).authenticatedResponseStringWithRetries()
@@ -405,81 +210,7 @@ class VardaClient(
         }
     }
 
-    fun updateDecision(vardaDecisionId: Long, updatedDecision: VardaDecision): VardaDecisionResponse? {
-        logger.info { "Updating a decision to Varda (vardaId: $vardaDecisionId, body: $updatedDecision)" }
-
-        val (request, _, result) = fuel.put(getDecisionUrl(vardaDecisionId))
-            .jsonBody(objectMapper.writeValueAsString(updatedDecision)).authenticatedResponseStringWithRetries()
-
-        return when (result) {
-            is Result.Success -> {
-                logger.info { "Updating a decision to Varda succeeded (vardaId: $vardaDecisionId, body: $updatedDecision)" }
-                objectMapper.readValue(result.get())
-            }
-            is Result.Failure -> {
-                logRequestError(request, result.error)
-                null
-            }
-        }
-    }
-
-    fun deleteDecision(vardaDecisionId: Long): Boolean {
-        logger.info { "Deleting decision from Varda (id: $vardaDecisionId)" }
-
-        val (request, _, result) = fuel.delete(getDecisionUrl(vardaDecisionId))
-            .authenticatedResponseStringWithRetries()
-
-        return when (result) {
-            is Result.Success -> {
-                logger.info { "Deleting decision from Varda succeeded (id: $vardaDecisionId)" }
-                true
-            }
-            is Result.Failure -> {
-                logRequestError(request, result.error)
-                false
-            }
-        }
-    }
-
-    fun deleteDecisionV2(vardaDecisionId: Long): Boolean {
-        logger.info { "VardaUpdate: client deleting decision (id: $vardaDecisionId)" }
-
-        val (request, _, result) = fuel.delete(getDecisionUrl(vardaDecisionId))
-            .authenticatedResponseStringWithRetries()
-
-        return when (result) {
-            is Result.Success -> {
-                logger.info { "VardaUpdate: client successfully deleted decision (id: $vardaDecisionId)" }
-                true
-            }
-            is Result.Failure -> {
-                vardaError(request, result.error) { err ->
-                    "VardaUpdate: client failed to delete decision $vardaDecisionId: $err"
-                }
-            }
-        }
-    }
-
-    fun createPlacement(newPlacement: VardaPlacement): VardaPlacementResponse? {
-        logger.info { "Creating a new placement to Varda (body: $newPlacement)" }
-
-        val (request, _, result) = fuel.post(placementUrl)
-            .jsonBody(objectMapper.writeValueAsString(newPlacement))
-            .authenticatedResponseStringWithRetries()
-
-        return when (result) {
-            is Result.Success -> {
-                logger.info { "Creating a new placement to Varda succeeded (body: $newPlacement)" }
-                objectMapper.readValue(result.get())
-            }
-            is Result.Failure -> {
-                logRequestError(request, result.error)
-                null
-            }
-        }
-    }
-
-    fun createPlacementV2(newPlacement: VardaPlacement): VardaPlacementResponse {
+    fun createPlacement(newPlacement: VardaPlacement): VardaPlacementResponse {
         logger.info { "VardaUpdate: client sending new placement (body: $newPlacement)" }
         val (request, _, result) = fuel.post(placementUrl)
             .jsonBody(objectMapper.writeValueAsString(newPlacement))
@@ -498,47 +229,43 @@ class VardaClient(
         }
     }
 
-    fun updatePlacement(
-        vardaPlacementId: Long,
-        updatedPlacement: VardaPlacement
-    ): VardaPlacementResponse? {
-        logger.info { "Updating a placement to Varda (body: $updatedPlacement)" }
-
-        val (request, _, result) = fuel.put(getPlacementUrl(vardaPlacementId))
-            .jsonBody(objectMapper.writeValueAsString(updatedPlacement))
-            .authenticatedResponseStringWithRetries()
+    fun createFeeData(feeData: VardaFeeData): VardaFeeDataResponse {
+        logger.info { "VardaUpdate: client sending fee data for child ${feeData.lapsi}" }
+        val (request, _, result) = fuel.post(feeDataUrl)
+            .jsonBody(objectMapper.writeValueAsString(feeData)).authenticatedResponseStringWithRetries()
 
         return when (result) {
             is Result.Success -> {
-                logger.info { "Updating a placement to Varda succeeded (body: $updatedPlacement)" }
+                logger.info { "VardaUpdate: client successfully sent fee data for child ${feeData.lapsi}" }
                 objectMapper.readValue(result.get())
             }
             is Result.Failure -> {
-                logRequestError(request, result.error)
-                null
+                vardaError(request, result.error) { err ->
+                    "VardaUpdate: client failed to send fee data for child ${feeData.lapsi}: ${err.errorCode}, $err"
+                }
+            }
+        }
+    }
+
+    fun deleteFeeData(vardaId: Long): Boolean {
+        logger.info { "VardaUpdate: client deleting fee data $vardaId" }
+
+        val (request, _, result) = fuel.delete("$feeDataUrl$vardaId/").authenticatedResponseStringWithRetries()
+
+        return when (result) {
+            is Result.Success -> {
+                logger.info { "VardaUpdate: client successfully deleted fee data $vardaId" }
+                true
+            }
+            is Result.Failure -> {
+                vardaError(request, result.error) { err ->
+                    "VardaUpdate: client failed to delete fee data $vardaId: $err"
+                }
             }
         }
     }
 
     fun deletePlacement(vardaPlacementId: Long): Boolean {
-        logger.info { "Deleting placement from Varda (id: $vardaPlacementId)" }
-
-        val (request, _, result) = fuel.delete(getPlacementUrl(vardaPlacementId))
-            .authenticatedResponseStringWithRetries()
-
-        return when (result) {
-            is Result.Success -> {
-                logger.info { "Deleting placement from Varda succeeded (id: $vardaPlacementId)" }
-                true
-            }
-            is Result.Failure -> {
-                logRequestError(request, result.error)
-                false
-            }
-        }
-    }
-
-    fun deletePlacementV2(vardaPlacementId: Long): Boolean {
         logger.info { "VardaUpdate: client deleting placement (id: $vardaPlacementId)" }
 
         val (request, _, result) = fuel.delete(getPlacementUrl(vardaPlacementId))
@@ -557,20 +284,81 @@ class VardaClient(
         }
     }
 
-    fun deleteChild(vardaChildId: Long): Boolean {
-        logger.info { "VardaUpdate: Deleting child (id: $vardaChildId)" }
+    fun deleteDecision(vardaDecisionId: Long): Boolean {
+        logger.info { "VardaUpdate: client deleting decision (id: $vardaDecisionId)" }
 
-        val (request, _, result) = fuel.delete(getChildUrl(vardaChildId))
+        val (request, _, result) = fuel.delete(getDecisionUrl(vardaDecisionId))
             .authenticatedResponseStringWithRetries()
 
         return when (result) {
             is Result.Success -> {
-                logger.info { "Deleting child from Varda succeeded (id: $vardaChildId)" }
+                logger.info { "VardaUpdate: client successfully deleted decision (id: $vardaDecisionId)" }
                 true
             }
             is Result.Failure -> {
-                logRequestError(request, result.error)
-                false
+                vardaError(request, result.error) { err ->
+                    "VardaUpdate: client failed to delete decision $vardaDecisionId: $err"
+                }
+            }
+        }
+    }
+
+    fun createUnit(unit: VardaUnitRequest): VardaUnitResponse {
+        logger.info { "VardaUpdate: client sending new unit ${unit.nimi}" }
+        val (request, _, result) = fuel.post(unitUrl)
+            .jsonBody(objectMapper.writeValueAsString(unit))
+            .authenticatedResponseStringWithRetries()
+
+        return when (result) {
+            is Result.Success -> {
+                logger.info { "VardaUpdate: client successfully sent new unit ${unit.nimi}" }
+                objectMapper.readValue(result.get())
+            }
+            is Result.Failure -> {
+                vardaError(request, result.error) { err ->
+                    "VardaUpdate: client failed to send unit ${unit.nimi}: $err"
+                }
+            }
+        }
+    }
+
+    fun updateUnit(unit: VardaUnitRequest): VardaUnitResponse {
+        logger.info { "VardaUpdate: client updating unit ${unit.nimi}" }
+        val url = "$unitUrl${unit.id}/"
+        val (request, _, result) = fuel.put(url)
+            .jsonBody(objectMapper.writeValueAsString(unit))
+            .authenticatedResponseStringWithRetries()
+
+        return when (result) {
+            is Result.Success -> {
+                logger.info { "VardaUpdate: client successfully updated unit ${unit.nimi}" }
+                objectMapper.readValue(result.get())
+            }
+            is Result.Failure -> {
+                vardaError(request, result.error) { err ->
+                    "VardaUpdate: client failed to update unit ${unit.nimi}: $err"
+                }
+            }
+        }
+    }
+
+    fun updateOrganizer(organizer: VardaUpdateOrganizer): Boolean {
+        logger.info { "VardaUpdate: client updating organizer" }
+
+        val (request, _, result) = fuel.put("$organizerUrl${organizer.vardaOrganizerId}")
+            .header(Headers.CONTENT_TYPE, "application/json")
+            .jsonBody(objectMapper.writeValueAsString(organizer))
+            .authenticatedResponseStringWithRetries()
+
+        return when (result) {
+            is Result.Success -> {
+                logger.info { "VardaUpdate: client successfully updated organizer" }
+                true
+            }
+            is Result.Failure -> {
+                vardaError(request, result.error) { err ->
+                    "VardaUpdate: client failed to update organizer: $err"
+                }
             }
         }
     }
@@ -598,19 +386,6 @@ class VardaClient(
         return getAllPages("$decisionUrl?lapsi=$vardaChildId") {
             objectMapper.readValue<PaginatedResponse<VardaResultId>>(it)
         }.map { it.id }
-    }
-
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    data class DecisionPeriod(
-        val id: Long,
-        val alkamis_pvm: LocalDate,
-        val paattymis_pvm: LocalDate
-    )
-
-    fun getChildDecisions(vardaChildId: Long): List<DecisionPeriod> {
-        return getAllPages("$childUrl$vardaChildId/varhaiskasvatuspaatokset/") {
-            objectMapper.readValue<PaginatedResponse<DecisionPeriod>>(it)
-        }
     }
 
     data class PaginatedResponse<T>(
@@ -642,9 +417,7 @@ class VardaClient(
             }
         }
 
-        val result = fetchNext(listOf(), initialUrl)
-
-        return result
+        return fetchNext(listOf(), initialUrl)
     }
 
     /**
@@ -679,38 +452,4 @@ class VardaClient(
                     }
                 }
         }
-}
-
-private fun logRequestError(request: Request, error: FuelError) {
-    val meta = mapOf(
-        "method" to request.method,
-        "url" to request.url,
-        "body" to request.body.asString("application/json"),
-        "errorMessage" to error.errorData.decodeToString()
-    )
-    logger.error(error, meta) { "Varda request failed to ${request.url}, status ${error.response.statusCode}" }
-}
-
-fun convertToVardaChildRequest(evakaPersonId: UUID, payload: VardaChildPayload): VardaChildRequest {
-    return VardaChildRequest(
-        id = evakaPersonId,
-        henkilo = payload.personUrl,
-        henkilo_oid = payload.personOid,
-        vakatoimija_oid = payload.organizerOid,
-        oma_organisaatio_oid = null,
-        paos_organisaatio_oid = null,
-        lahdejarjestelma = payload.sourceSystem
-    )
-}
-
-fun convertToVardaChildRequest(evakaPersonId: UUID, payload: VardaPaosChildPayload): VardaChildRequest {
-    return VardaChildRequest(
-        id = evakaPersonId,
-        henkilo = payload.personUrl,
-        henkilo_oid = payload.personOid,
-        vakatoimija_oid = null,
-        oma_organisaatio_oid = payload.organizerOid,
-        paos_organisaatio_oid = payload.paosOrganizationOid,
-        lahdejarjestelma = payload.sourceSystem
-    )
 }
